Add if/else statement to deleteoldestbackup.ps1 based on boolean of checkenoughspace and checkhowmanybackup

+++ deleteoldestbackup.ps1
@@ -1,4 +1,9 @@
 function deleteoldestbackup {
-    $allFolders | Sort-Object CreationTime -Descending | Select-Object -Skip 1 | Remove-Item -Recurse -Force
-    Write-Output 'Deletion completed. Creating new backup'
+    if ((checkenoughspace) -eq $false -And (checkhowmanybackup -eq $true)) {
+        $allFolders | Sort-Object CreationTime -Descending | Select-Object -Skip 1 | Remove-Item -Recurse -Force
+        Write-Output 'Deletion completed. Creating new backup'
+    } else {
+        Write-Output 'The script has encountered error(s). Please check code.'
+        exit
+    }
 }
